move factory functions in their own file
